Add upcoming notables and visual coverage
