+++ fredagskoll-frontend/playwright.config.ts
@@ -1,0 +1,35 @@
+import { defineConfig, devices } from '@playwright/test';
+
+export default defineConfig({
+  testDir: './tests',
+  timeout: 30_000,
+  expect: {
+    toHaveScreenshot: {
+      maxDiffPixelRatio: 0.02,
+    },
+  },
+  outputDir: './output/playwright/test-results',
+  snapshotPathTemplate:
+    '{testDir}/__screenshots__/{testFilePath}/{arg}{ext}',
+  use: {
+    baseURL: 'http://127.0.0.1:3000',
+    trace: 'retain-on-failure',
+  },
+  projects: [
+    {
+      name: 'chromium',
+      use: {
+        ...devices['Desktop Chrome'],
+        viewport: { width: 1440, height: 1280 },
+      },
+    },
+  ],
+  webServer: {
+    command: 'npm start',
+    url: 'http://127.0.0.1:3000',
+    reuseExistingServer: true,
+    stdout: 'ignore',
+    stderr: 'pipe',
+    timeout: 120_000,
+  },
+});
